+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Versioning](http://semver.org/spec/v2.0.0.html) except to the first release.
 
 - Implemented all box.schema.user operations requests and sugar interface (#426).
 - Implemented box.session.su request and sugar interface only for current session granting (#426).
+- Defined `ErrConcurrentSchemaUpdate` constant for "concurrent schema update" error.
+  Now you can check this error with `errors.Is(err, tarantool.ErrConcurrentSchemaUpdate)`.
 
 ### Changed